Add pow() with compile-time exponent
This is basically a copy of the regular pow() implementation.

Only a signed version is offered instead of both a signed and unsigned
version because having both a signed long long and unsigned long long
version will result in ambiguous calls for common uses like pow<2>(x).
This does mean that calculation of ludicrously large exponents isn't
supported, but hopefully such cases are exceptionally rare.