Fix incorrect trig function output for scaled dimensionless types
This is a port of the analogous changes made to the v2.x branch.

Similar to the previous commit, but for certain trigonometric functions.
I don't think that explicitly calling to<>() is strictly necessary for
the other trig functions since they already use convert<>() to
normalize, so I don't think normalization through to<>() is necessary.
I have not tested it, though.

Not entirely sure the modifications to the test cases are the best way
to make the desired changes. The idea is to use the same inputs/outputs,
but to change the unit type to have a conversion factor not equal to 1.